Move download file function to actions/content.js
